Fixing selector issue
So for the orginal SASS that was here &.background-image is the same as writing .menu-item.background-image, which is looking for an element with the classes
<div class='menu-item background-image'></div>

The actual code you want to target is a div inside of the element with the .menu-item class so in normal css you would put that as .menu-item .background-image with a space between the two classes.  Same applies here, just need to add a space between the & and the .background-image and it should hopefully work.